feat & refactor: api request to prediction server (dummy), separate steps into files.
